+++ src/nts_processing/hierachical_bayesian_model/hierachical_bayesian_run_code.py
@@ -8,16 +8,14 @@ Original author: Adil Zaheer
 # pylint: enable=import-error,wrong-import-position
 import warnings
 
-import numpy as np
-
-from src.nts_processing.AZ_code_MTS_model.process_cb_functions import (process_cb_data_atkins_method,
-                                                                       process_data_for_hierarchical_bayesian_model,
-                                                                       generate_priors,
-                                                                       predict_model_hierarchical_bayesian_model,
-                                                                       load_model_and_parameters,
-                                                                       save_model_and_parameters,
-                                                                       evaluate_model,
-                                                                       process_cb_data_tfn_method)
+from src.nts_processing.processing_classified_build.process_cb_functions import (process_cb_data_atkins_method,
+                                                                                 process_data_for_hierarchical_bayesian_model,
+                                                                                 generate_priors,
+                                                                                 predict_model_hierarchical_bayesian_model,
+                                                                                 load_model_and_parameters,
+                                                                                 save_model_and_parameters,
+                                                                                 evaluate_model,
+                                                                                 process_cb_data_tfn_method)
 
 warnings.filterwarnings("ignore")
