+++ src/nts_processing/production_weights_and_MTS/prod_weight_mts_functions.py
@@ -19,7 +19,7 @@ from sklearn.ensemble import GradientBoostingRegressor
 from sklearn.metrics import mean_squared_error, r2_score, mean_absolute_error, make_scorer
 from caf.ml.functions import data_pipeline_functions as dpf
 from caf.ml.functions import process_data_functions as pdf
-from src.nts_processing.AZ_code_MTS_model import process_cb_functions as mts_process
+from src.nts_processing.processing_classified_build import process_cb_functions as mts_process
 
 
 class TripRate:
